Move out most of the cross-version tests from integ-test
