+++ platforms/extensibility/plugin-use/src/crossVersionTest/groovy/org/gradle/integtests/compatibility/AbstractPropertyUpgradesBinaryCompatibilityCrossVersionSpec.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 the original author or authors.
+ * Copyright 2024 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradle.integtests
+
+package org.gradle.integtests.compatibility
+
+
 
 import org.gradle.configuration.DefaultImportsReader
 import org.gradle.integtests.fixtures.CrossVersionIntegrationSpec
