+++ platforms/extensibility/plugin-use/src/crossVersionTest/groovy/org/gradle/integtests/compatibility/AbstractTaskSubclassingBinaryCompatibilityCrossVersionSpec.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2012 the original author or authors.
+ * Copyright 2024 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradle.integtests
+
+package org.gradle.integtests.compatibility
+
+
 
 import org.gradle.api.DefaultTask
 import org.gradle.api.internal.ConventionTask
@@ -51,33 +54,33 @@ import org.gradle.util.GradleVersion
 abstract class AbstractTaskSubclassingBinaryCompatibilityCrossVersionSpec extends CrossVersionIntegrationSpec {
     protected void prepareSubclassingTest(GradleVersion targetVersion) {
         def taskClasses = [
-            DefaultTask,
-            SourceTask,
-            ConventionTask,
-            Copy,
-            Sync,
-            Zip,
-            org.gradle.api.tasks.bundling.Jar,
-            Tar,
-            War,
-            JavaCompile,
-            GroovyCompile,
-            ScalaCompile,
-            Test,
-            CodeNarc,
-            Checkstyle,
-            Ear,
-            Pmd,
-            Sign,
-            org.gradle.api.tasks.application.CreateStartScripts,
-            GenerateEclipseJdt,
-            GenerateEclipseClasspath,
-            GenerateEclipseProject,
-            GenerateEclipseWtpComponent,
-            GenerateEclipseWtpFacet,
-            GenerateIdeaModule,
-            GenerateIdeaWorkspace,
-            GenerateIdeaProject
+                DefaultTask,
+                SourceTask,
+                ConventionTask,
+                Copy,
+                Sync,
+                Zip,
+                org.gradle.api.tasks.bundling.Jar,
+                Tar,
+                War,
+                JavaCompile,
+                GroovyCompile,
+                ScalaCompile,
+                Test,
+                CodeNarc,
+                Checkstyle,
+                Ear,
+                Pmd,
+                Sign,
+                org.gradle.api.tasks.application.CreateStartScripts,
+                GenerateEclipseJdt,
+                GenerateEclipseClasspath,
+                GenerateEclipseProject,
+                GenerateEclipseWtpComponent,
+                GenerateEclipseWtpFacet,
+                GenerateIdeaModule,
+                GenerateIdeaWorkspace,
+                GenerateIdeaProject
         ]
 
         // Task types added after 1.0
